Updated changelog.
