+++ arangod/RocksDBEngine/RocksDBRestReplicationHandler.cpp
@@ -73,18 +73,12 @@ using namespace arangodb::basics;
 using namespace arangodb::rest;
 using namespace arangodb::rocksutils;
 
-static int restoreDataParser(char const* ptr, char const* pos,
-                             std::string const& invalidMsg, bool useRevision,
-                             std::string& errorMsg, std::string& key,
-                             VPackBuilder& builder, VPackSlice& doc,
-                             TRI_replication_operation_e& type);
-
 uint64_t const RocksDBRestReplicationHandler::_defaultChunkSize = 128 * 1024;
 uint64_t const RocksDBRestReplicationHandler::_maxChunkSize = 128 * 1024 * 1024;
 
 RocksDBRestReplicationHandler::RocksDBRestReplicationHandler(
     GeneralRequest* request, GeneralResponse* response)
-    : RestVocbaseBaseHandler(request, response),
+    : RestReplicationHandler(request, response),
       _manager(globalRocksEngine()->replicationManager()) {}
 
 RocksDBRestReplicationHandler::~RocksDBRestReplicationHandler() {}
@@ -310,21 +304,6 @@ BAD_CALL:
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-/// @brief creates an error if called on a coordinator server
-////////////////////////////////////////////////////////////////////////////////
-
-bool RocksDBRestReplicationHandler::isCoordinatorError() {
-  if (_vocbase->type() == TRI_VOCBASE_TYPE_COORDINATOR) {
-    generateError(rest::ResponseCode::NOT_IMPLEMENTED,
-                  TRI_ERROR_CLUSTER_UNSUPPORTED,
-                  "replication API is not supported on a coordinator");
-    return true;
-  }
-
-  return false;
-}
-
-////////////////////////////////////////////////////////////////////////////////
 /// @brief was docuBlock JSF_get_api_replication_logger_return_state
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -465,128 +444,6 @@ void RocksDBRestReplicationHandler::handleCommandBatch() {
   // we get here if anything above is invalid
   generateError(rest::ResponseCode::METHOD_NOT_ALLOWED,
                 TRI_ERROR_HTTP_METHOD_NOT_ALLOWED);
-}
-
-////////////////////////////////////////////////////////////////////////////////
-/// @brief forward a command in the coordinator case
-////////////////////////////////////////////////////////////////////////////////
-
-void RocksDBRestReplicationHandler::handleTrampolineCoordinator() {
-  bool useVst = false;
-  if (_request->transportType() == Endpoint::TransportType::VST) {
-    useVst = true;
-  }
-  if (_request == nullptr) {
-    THROW_ARANGO_EXCEPTION_MESSAGE(TRI_ERROR_INTERNAL, "invalid request");
-  }
-
-  // First check the DBserver component of the body json:
-  ServerID DBserver = _request->value("DBserver");
-
-  if (DBserver.empty()) {
-    generateError(rest::ResponseCode::BAD, TRI_ERROR_HTTP_BAD_PARAMETER,
-                  "need \"DBserver\" parameter");
-    return;
-  }
-
-  std::string const& dbname = _request->databaseName();
-
-  auto headers = std::make_shared<std::unordered_map<std::string, std::string>>(
-      arangodb::getForwardableRequestHeaders(_request.get()));
-  std::unordered_map<std::string, std::string> values = _request->values();
-  std::string params;
-
-  for (auto const& i : values) {
-    if (i.first != "DBserver") {
-      if (params.empty()) {
-        params.push_back('?');
-      } else {
-        params.push_back('&');
-      }
-      params.append(StringUtils::urlEncode(i.first));
-      params.push_back('=');
-      params.append(StringUtils::urlEncode(i.second));
-    }
-  }
-
-  // Set a few variables needed for our work:
-  auto cc = ClusterComm::instance();
-  if (cc == nullptr) {
-    // nullptr happens only during controlled shutdown
-    generateError(rest::ResponseCode::BAD, TRI_ERROR_SHUTTING_DOWN,
-                  "shutting down server");
-    return;
-  }
-
-  std::unique_ptr<ClusterCommResult> res;
-  if (!useVst) {
-    HttpRequest* httpRequest = dynamic_cast<HttpRequest*>(_request.get());
-    if (httpRequest == nullptr) {
-      THROW_ARANGO_EXCEPTION_MESSAGE(TRI_ERROR_INTERNAL,
-                                     "invalid request type");
-    }
-
-    // Send a synchronous request to that shard using ClusterComm:
-    res = cc->syncRequest("", TRI_NewTickServer(), "server:" + DBserver,
-                          _request->requestType(),
-                          "/_db/" + StringUtils::urlEncode(dbname) +
-                              _request->requestPath() + params,
-                          httpRequest->body(), *headers, 300.0);
-  } else {
-    // do we need to handle multiple payloads here - TODO
-    // here we switch vorm vst to http?!
-    // i am not allowed to change cluster comm!
-    res = cc->syncRequest("", TRI_NewTickServer(), "server:" + DBserver,
-                          _request->requestType(),
-                          "/_db/" + StringUtils::urlEncode(dbname) +
-                              _request->requestPath() + params,
-                          _request->payload().toJson(), *headers, 300.0);
-  }
-
-  if (res->status == CL_COMM_TIMEOUT) {
-    // No reply, we give up:
-    generateError(rest::ResponseCode::BAD, TRI_ERROR_CLUSTER_TIMEOUT,
-                  "timeout within cluster");
-    return;
-  }
-  if (res->status == CL_COMM_BACKEND_UNAVAILABLE) {
-    // there is no result
-    generateError(rest::ResponseCode::BAD, TRI_ERROR_CLUSTER_CONNECTION_LOST,
-                  "lost connection within cluster");
-    return;
-  }
-  if (res->status == CL_COMM_ERROR) {
-    // This could be a broken connection or an Http error:
-    TRI_ASSERT(nullptr != res->result && res->result->isComplete());
-    // In this case a proper HTTP error was reported by the DBserver,
-    // we simply forward the result.
-    // We intentionally fall through here.
-  }
-
-  bool dummy;
-  resetResponse(
-      static_cast<rest::ResponseCode>(res->result->getHttpReturnCode()));
-
-  _response->setContentType(
-      res->result->getHeaderField(StaticStrings::ContentTypeHeader, dummy));
-
-  if (!useVst) {
-    HttpResponse* httpResponse = dynamic_cast<HttpResponse*>(_response.get());
-    if (_response == nullptr) {
-      THROW_ARANGO_EXCEPTION_MESSAGE(TRI_ERROR_INTERNAL,
-                                     "invalid response type");
-    }
-    httpResponse->body().swap(&(res->result->getBody()));
-  } else {
-    // TODO copy all payloads
-    VPackSlice slice = res->result->getBodyVelocyPack()->slice();
-    _response->setPayload(slice, true);  // do we need to generate the body?!
-  }
-
-  auto const& resultHeaders = res->result->getHeaderFields();
-  for (auto const& it : resultHeaders) {
-    _response->setHeader(it.first, it.second);
-  }
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -840,41 +697,6 @@ void RocksDBRestReplicationHandler::handleCommandInventory() {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-/// @brief was docuBlock JSF_get_api_replication_cluster_inventory
-////////////////////////////////////////////////////////////////////////////////
-
-void RocksDBRestReplicationHandler::handleCommandClusterInventory() {
-  std::string const& dbName = _request->databaseName();
-  bool found;
-  bool includeSystem = true;
-
-  std::string const& value = _request->value("includeSystem", found);
-
-  if (found) {
-    includeSystem = StringUtils::boolean(value);
-  }
-
-  ClusterInfo* ci = ClusterInfo::instance();
-  std::vector<std::shared_ptr<LogicalCollection>> cols =
-      ci->getCollections(dbName);
-
-  VPackBuilder resultBuilder;
-  resultBuilder.openObject();
-  resultBuilder.add(VPackValue("collections"));
-  resultBuilder.openArray();
-  for (auto const& c : cols) {
-    c->toVelocyPackForClusterInventory(resultBuilder, includeSystem);
-  }
-  resultBuilder.close();  // collections
-  TRI_voc_tick_t tick = TRI_CurrentTickServer();
-  auto tickString = std::to_string(tick);
-  resultBuilder.add("tick", VPackValue(tickString));
-  resultBuilder.add("state", VPackValue("unused"));
-  resultBuilder.close();  // base
-  generateResult(rest::ResponseCode::OK, resultBuilder.slice());
-}
-
-////////////////////////////////////////////////////////////////////////////////
 /// @brief restores the structure of a collection TODO MOVE
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -960,92 +782,6 @@ void RocksDBRestReplicationHandler::handleCommandRestoreCollection() {
   result.add("result", VPackValue(true));
   result.close();
   generateResult(rest::ResponseCode::OK, result.slice());
-}
-
-////////////////////////////////////////////////////////////////////////////////
-/// @brief restores the indexes of a collection TODO MOVE
-////////////////////////////////////////////////////////////////////////////////
-
-void RocksDBRestReplicationHandler::handleCommandRestoreIndexes() {
-  std::shared_ptr<VPackBuilder> parsedRequest;
-
-  try {
-    parsedRequest = _request->toVelocyPackBuilderPtr();
-  } catch (...) {
-    generateError(rest::ResponseCode::BAD, TRI_ERROR_HTTP_BAD_PARAMETER,
-                  "invalid JSON");
-    return;
-  }
-  VPackSlice const slice = parsedRequest->slice();
-
-  bool found;
-  bool force = false;
-  std::string const& value = _request->value("force", found);
-
-  if (found) {
-    force = StringUtils::boolean(value);
-  }
-
-  std::string errorMsg;
-  int res;
-  if (ServerState::instance()->isCoordinator()) {
-    res = processRestoreIndexesCoordinator(slice, force, errorMsg);
-  } else {
-    res = processRestoreIndexes(slice, force, errorMsg);
-  }
-
-  if (res != TRI_ERROR_NO_ERROR) {
-    THROW_ARANGO_EXCEPTION(res);
-  }
-
-  VPackBuilder result;
-  result.openObject();
-  result.add("result", VPackValue(true));
-  result.close();
-  generateResult(rest::ResponseCode::OK, result.slice());
-}
-
-void RocksDBRestReplicationHandler::handleCommandRestoreData() {
-  std::string const& colName = _request->value("collection");
-
-  if (colName.empty()) {
-    generateError(rest::ResponseCode::BAD, TRI_ERROR_HTTP_BAD_PARAMETER,
-                  "invalid collection parameter, not given");
-    return;
-  }
-
-  bool recycleIds = false;
-  std::string const& value2 = _request->value("recycleIds");
-
-  if (!value2.empty()) {
-    recycleIds = StringUtils::boolean(value2);
-  }
-
-  bool force = false;
-  std::string const& value3 = _request->value("force");
-
-  if (!value3.empty()) {
-    force = StringUtils::boolean(value3);
-  }
-
-  std::string errorMsg;
-
-  int res = processRestoreData(colName, recycleIds, force, errorMsg);
-
-  if (res != TRI_ERROR_NO_ERROR) {
-    if (errorMsg.empty()) {
-      generateError(GeneralResponse::responseCode(res), res);
-    } else {
-      generateError(GeneralResponse::responseCode(res), res,
-                    std::string(TRI_errno_string(res)) + ": " + errorMsg);
-    }
-  } else {
-    VPackBuilder result;
-    result.add(VPackValue(VPackValueType::Object));
-    result.add("result", VPackValue(true));
-    result.close();
-    generateResult(rest::ResponseCode::OK, result.slice());
-  }
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -1352,183 +1088,6 @@ void RocksDBRestReplicationHandler::handleCommandDump() {
 
   // avoid double freeing
   TRI_StealStringBuffer(dump.stringBuffer());
-}
-
-////////////////////////////////////////////////////////////////////////////////
-/// @brief was docuBlock JSF_put_api_replication_makeSlave
-////////////////////////////////////////////////////////////////////////////////
-
-void RocksDBRestReplicationHandler::handleCommandMakeSlave() {
-  // NOTE: Complete copy of MMFiles
-  bool success;
-  std::shared_ptr<VPackBuilder> parsedBody = parseVelocyPackBody(success);
-  if (!success) {
-    // error already created
-    return;
-  }
-  VPackSlice const body = parsedBody->slice();
-
-  std::string const endpoint =
-      VelocyPackHelper::getStringValue(body, "endpoint", "");
-
-  if (endpoint.empty()) {
-    generateError(rest::ResponseCode::BAD, TRI_ERROR_HTTP_BAD_PARAMETER,
-                  "<endpoint> must be a valid endpoint");
-    return;
-  }
-
-  std::string const database =
-      VelocyPackHelper::getStringValue(body, "database", _vocbase->name());
-  std::string const username =
-      VelocyPackHelper::getStringValue(body, "username", "");
-  std::string const password =
-      VelocyPackHelper::getStringValue(body, "password", "");
-  std::string const jwt = VelocyPackHelper::getStringValue(body, "jwt", "");
-  std::string const restrictType =
-      VelocyPackHelper::getStringValue(body, "restrictType", "");
-
-  // initialize some defaults to copy from
-  TRI_replication_applier_configuration_t defaults;
-
-  // initialize target configuration
-  TRI_replication_applier_configuration_t config;
-
-  config._endpoint = endpoint;
-  config._database = database;
-  config._username = username;
-  config._password = password;
-  config._jwt = jwt;
-  config._includeSystem =
-      VelocyPackHelper::getBooleanValue(body, "includeSystem", true);
-  config._requestTimeout = VelocyPackHelper::getNumericValue<double>(
-      body, "requestTimeout", defaults._requestTimeout);
-  config._connectTimeout = VelocyPackHelper::getNumericValue<double>(
-      body, "connectTimeout", defaults._connectTimeout);
-  config._ignoreErrors = VelocyPackHelper::getNumericValue<uint64_t>(
-      body, "ignoreErrors", defaults._ignoreErrors);
-  config._maxConnectRetries = VelocyPackHelper::getNumericValue<uint64_t>(
-      body, "maxConnectRetries", defaults._maxConnectRetries);
-  config._sslProtocol = VelocyPackHelper::getNumericValue<uint32_t>(
-      body, "sslProtocol", defaults._sslProtocol);
-  config._chunkSize = VelocyPackHelper::getNumericValue<uint64_t>(
-      body, "chunkSize", defaults._chunkSize);
-  config._autoStart = true;
-  config._adaptivePolling = VelocyPackHelper::getBooleanValue(
-      body, "adaptivePolling", defaults._adaptivePolling);
-  config._autoResync = VelocyPackHelper::getBooleanValue(body, "autoResync",
-                                                         defaults._autoResync);
-  config._verbose =
-      VelocyPackHelper::getBooleanValue(body, "verbose", defaults._verbose);
-  config._incremental = VelocyPackHelper::getBooleanValue(
-      body, "incremental", defaults._incremental);
-  config._useCollectionId = VelocyPackHelper::getBooleanValue(
-      body, "useCollectionId", defaults._useCollectionId);
-  config._requireFromPresent = VelocyPackHelper::getBooleanValue(
-      body, "requireFromPresent", defaults._requireFromPresent);
-  config._restrictType = VelocyPackHelper::getStringValue(
-      body, "restrictType", defaults._restrictType);
-  config._connectionRetryWaitTime = static_cast<uint64_t>(
-      1000.0 * 1000.0 *
-      VelocyPackHelper::getNumericValue<double>(
-          body, "connectionRetryWaitTime",
-          static_cast<double>(defaults._connectionRetryWaitTime) /
-              (1000.0 * 1000.0)));
-  config._initialSyncMaxWaitTime = static_cast<uint64_t>(
-      1000.0 * 1000.0 *
-      VelocyPackHelper::getNumericValue<double>(
-          body, "initialSyncMaxWaitTime",
-          static_cast<double>(defaults._initialSyncMaxWaitTime) /
-              (1000.0 * 1000.0)));
-  config._idleMinWaitTime = static_cast<uint64_t>(
-      1000.0 * 1000.0 *
-      VelocyPackHelper::getNumericValue<double>(
-          body, "idleMinWaitTime",
-          static_cast<double>(defaults._idleMinWaitTime) / (1000.0 * 1000.0)));
-  config._idleMaxWaitTime = static_cast<uint64_t>(
-      1000.0 * 1000.0 *
-      VelocyPackHelper::getNumericValue<double>(
-          body, "idleMaxWaitTime",
-          static_cast<double>(defaults._idleMaxWaitTime) / (1000.0 * 1000.0)));
-  config._autoResyncRetries = VelocyPackHelper::getNumericValue<uint64_t>(
-      body, "autoResyncRetries", defaults._autoResyncRetries);
-
-  VPackSlice const restriction = body.get("restrictCollections");
-
-  if (restriction.isArray()) {
-    VPackValueLength const n = restriction.length();
-
-    for (VPackValueLength i = 0; i < n; ++i) {
-      VPackSlice const cname = restriction.at(i);
-      if (cname.isString()) {
-        config._restrictCollections.emplace(cname.copyString(), true);
-      }
-    }
-  }
-
-  // now the configuration is complete
-
-  if ((restrictType.empty() && !config._restrictCollections.empty()) ||
-      (!restrictType.empty() && config._restrictCollections.empty()) ||
-      (!restrictType.empty() && restrictType != "include" &&
-       restrictType != "exclude")) {
-    generateError(rest::ResponseCode::BAD, TRI_ERROR_HTTP_BAD_PARAMETER,
-                  "invalid value for <restrictCollections> or <restrictType>");
-    return;
-  }
-
-  // forget about any existing replication applier configuration
-  int res = _vocbase->replicationApplier()->forget();
-
-  if (res != TRI_ERROR_NO_ERROR) {
-    THROW_ARANGO_EXCEPTION(res);
-  }
-
-  // start initial synchronization
-  TRI_voc_tick_t lastLogTick = 0;
-  TRI_voc_tick_t barrierId = 0;
-  std::string errorMsg = "";
-  {
-    InitialSyncer syncer(_vocbase, &config, config._restrictCollections,
-                         restrictType, false, false);
-
-    res = TRI_ERROR_NO_ERROR;
-
-    try {
-      res = syncer.run(errorMsg, false);
-
-      // steal the barrier from the syncer
-      barrierId = syncer.stealBarrier();
-    } catch (...) {
-      errorMsg = "caught an exception";
-      res = TRI_ERROR_INTERNAL;
-    }
-
-    lastLogTick = syncer.getLastLogTick();
-  }
-
-  if (res != TRI_ERROR_NO_ERROR) {
-    THROW_ARANGO_EXCEPTION_MESSAGE(res, errorMsg);
-    return;
-  }
-
-  res =
-      TRI_ConfigureReplicationApplier(_vocbase->replicationApplier(), &config);
-
-  if (res != TRI_ERROR_NO_ERROR) {
-    THROW_ARANGO_EXCEPTION(res);
-    return;
-  }
-
-  res = _vocbase->replicationApplier()->start(lastLogTick, true, barrierId);
-
-  if (res != TRI_ERROR_NO_ERROR) {
-    THROW_ARANGO_EXCEPTION(res);
-    return;
-  }
-
-  std::shared_ptr<VPackBuilder> result =
-      _vocbase->replicationApplier()->toVelocyPack();
-  generateResult(rest::ResponseCode::OK, result->slice());
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -2815,364 +2374,6 @@ int RocksDBRestReplicationHandler::processRestoreIndexesCoordinator(
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-/// @brief restores the data of a collection TODO MOVE
-////////////////////////////////////////////////////////////////////////////////
-
-int RocksDBRestReplicationHandler::processRestoreDataBatch(
-    transaction::Methods& trx, std::string const& collectionName,
-    bool useRevision, bool force, std::string& errorMsg) {
-  std::string const invalidMsg =
-      "received invalid JSON data for collection " + collectionName;
-
-  VPackBuilder builder;
-
-  HttpRequest* httpRequest = dynamic_cast<HttpRequest*>(_request.get());
-  if (httpRequest == nullptr) {
-    THROW_ARANGO_EXCEPTION_MESSAGE(TRI_ERROR_INTERNAL, "invalid request type");
-  }
-
-  std::string const& bodyStr = httpRequest->body();
-  char const* ptr = bodyStr.c_str();
-  char const* end = ptr + bodyStr.size();
-
-  VPackBuilder allMarkers;
-  VPackValueLength currentPos = 0;
-  std::unordered_map<std::string, VPackValueLength> latest;
-
-  // First parse and collect all markers, we assemble everything in one
-  // large builder holding an array. We keep for each key the latest
-  // entry.
-
-  {
-    VPackArrayBuilder guard(&allMarkers);
-    std::string key;
-    while (ptr < end) {
-      char const* pos = strchr(ptr, '\n');
-
-      if (pos == nullptr) {
-        pos = end;
-      } else {
-        *((char*)pos) = '\0';
-      }
-
-      if (pos - ptr > 1) {
-        // found something
-        key.clear();
-        VPackSlice doc;
-        TRI_replication_operation_e type = REPLICATION_INVALID;
-
-        int res = restoreDataParser(ptr, pos, invalidMsg, useRevision, errorMsg,
-                                    key, builder, doc, type);
-        if (res != TRI_ERROR_NO_ERROR) {
-          return res;
-        }
-
-        // Put into array of all parsed markers:
-        allMarkers.add(builder.slice());
-        auto it = latest.find(key);
-        if (it != latest.end()) {
-          // Already found, overwrite:
-          it->second = currentPos;
-        } else {
-          latest.emplace(std::make_pair(key, currentPos));
-        }
-        ++currentPos;
-      }
-
-      ptr = pos + 1;
-    }
-  }
-
-  // First remove all keys of which the last marker we saw was a deletion
-  // marker:
-  VPackSlice allMarkersSlice = allMarkers.slice();
-  VPackBuilder oldBuilder;
-  {
-    VPackArrayBuilder guard(&oldBuilder);
-
-    for (auto const& p : latest) {
-      VPackSlice const marker = allMarkersSlice.at(p.second);
-      VPackSlice const typeSlice = marker.get("type");
-      TRI_replication_operation_e type = REPLICATION_INVALID;
-      if (typeSlice.isNumber()) {
-        int typeInt = typeSlice.getNumericValue<int>();
-        if (typeInt == 2301) {  // pre-3.0 type for edges
-          type = REPLICATION_MARKER_DOCUMENT;
-        } else {
-          type = static_cast<TRI_replication_operation_e>(typeInt);
-        }
-      }
-      if (type == REPLICATION_MARKER_REMOVE) {
-        oldBuilder.add(VPackValue(p.first));  // Add _key
-      } else if (type != REPLICATION_MARKER_DOCUMENT) {
-        errorMsg = "unexpected marker type " + StringUtils::itoa(type);
-        return TRI_ERROR_REPLICATION_UNEXPECTED_MARKER;
-      }
-    }
-  }
-
-  // Note that we ignore individual errors here, as long as the main
-  // operation did not fail. In particular, we intentionally ignore
-  // individual "DOCUMENT NOT FOUND" errors, because they can happen!
-  try {
-    OperationOptions options;
-    options.silent = true;
-    options.ignoreRevs = true;
-    options.isRestore = true;
-    options.waitForSync = false;
-    OperationResult opRes =
-        trx.remove(collectionName, oldBuilder.slice(), options);
-    if (!opRes.successful()) {
-      return opRes.code;
-    }
-  } catch (arangodb::basics::Exception const& ex) {
-    return ex.code();
-  } catch (...) {
-    return TRI_ERROR_INTERNAL;
-  }
-
-  // Now try to insert all keys for which the last marker was a document
-  // marker, note that these could still be replace markers!
-  builder.clear();
-  if (collectionName == "_users") {
-    VPackArrayBuilder guard(&builder);
-
-    for (auto const& p : latest) {
-      VPackSlice const marker = allMarkersSlice.at(p.second);
-      VPackSlice const typeSlice = marker.get("type");
-      TRI_replication_operation_e type = REPLICATION_INVALID;
-      if (typeSlice.isNumber()) {
-        int typeInt = typeSlice.getNumericValue<int>();
-        if (typeInt == 2301) {  // pre-3.0 type for edges
-          type = REPLICATION_MARKER_DOCUMENT;
-        } else {
-          type = static_cast<TRI_replication_operation_e>(typeInt);
-        }
-      }
-      if (type == REPLICATION_MARKER_DOCUMENT) {
-        VPackSlice const doc = marker.get("data");
-        TRI_ASSERT(doc.isObject());
-        // In the _users case we silently remove the _key value.
-        builder.openObject();
-        for (auto const& it : VPackObjectIterator(doc)) {
-          if (StringRef(it.key) != StaticStrings::KeyString && 
-              StringRef(it.key) != StaticStrings::IdString) {
-            builder.add(it.key);
-            builder.add(it.value);
-          }
-        }
-        builder.close();
-        builder.add(doc);
-      }
-    }
-  } else {
-    VPackArrayBuilder guard(&builder);
-
-    for (auto const& p : latest) {
-      VPackSlice const marker = allMarkersSlice.at(p.second);
-      VPackSlice const typeSlice = marker.get("type");
-      TRI_replication_operation_e type = REPLICATION_INVALID;
-      if (typeSlice.isNumber()) {
-        int typeInt = typeSlice.getNumericValue<int>();
-        if (typeInt == 2301) {  // pre-3.0 type for edges
-          type = REPLICATION_MARKER_DOCUMENT;
-        } else {
-          type = static_cast<TRI_replication_operation_e>(typeInt);
-        }
-      }
-      if (type == REPLICATION_MARKER_DOCUMENT) {
-        VPackSlice const doc = marker.get("data");
-        TRI_ASSERT(doc.isObject());
-        builder.add(doc);
-      }
-    }
-  }
-
-  VPackSlice requestSlice = builder.slice();
-  OperationResult opRes;
-  try {
-    OperationOptions options;
-    options.silent = false;
-    options.ignoreRevs = true;
-    options.isRestore = true;
-    options.waitForSync = false;
-    opRes = trx.insert(collectionName, requestSlice, options);
-    if (!opRes.successful()) {
-      return opRes.code;
-    }
-  } catch (arangodb::basics::Exception const& ex) {
-    return ex.code();
-  } catch (...) {
-    return TRI_ERROR_INTERNAL;
-  }
-
-  // Now go through the individual results and check each error, if it was
-  // TRI_ERROR_ARANGO_UNIQUE_CONSTRAINT_VIOLATED, then we have to call
-  // replace on the document:
-  VPackSlice resultSlice = opRes.slice();
-  VPackBuilder replBuilder;  // documents for replace operation
-  {
-    VPackArrayBuilder guard(&oldBuilder);
-    VPackArrayBuilder guard2(&replBuilder);
-    VPackArrayIterator itRequest(requestSlice);
-    VPackArrayIterator itResult(resultSlice);
-
-    while (itRequest.valid()) {
-      VPackSlice result = *itResult;
-      VPackSlice error = result.get("error");
-      if (error.isTrue()) {
-        error = result.get("errorNum");
-        if (error.isNumber()) {
-          int code = error.getNumericValue<int>();
-          if (code == TRI_ERROR_ARANGO_UNIQUE_CONSTRAINT_VIOLATED) {
-            replBuilder.add(*itRequest);
-          } else {
-            return code;
-          }
-        } else {
-          return TRI_ERROR_INTERNAL;
-        }
-      }
-      itRequest.next();
-      itResult.next();
-    }
-  }
-  try {
-    OperationOptions options;
-    options.silent = true;
-    options.ignoreRevs = true;
-    options.isRestore = true;
-    options.waitForSync = false;
-    opRes = trx.replace(collectionName, replBuilder.slice(), options);
-    if (!opRes.successful()) {
-      return opRes.code;
-    }
-  } catch (arangodb::basics::Exception const& ex) {
-    return ex.code();
-  } catch (...) {
-    return TRI_ERROR_INTERNAL;
-  }
-
-  return TRI_ERROR_NO_ERROR;
-}
-
-////////////////////////////////////////////////////////////////////////////////
-/// @brief restores the data of a collection TODO MOVE
-////////////////////////////////////////////////////////////////////////////////
-
-int RocksDBRestReplicationHandler::processRestoreData(
-    std::string const& colName, bool useRevision, bool force,
-    std::string& errorMsg) {
-
-  grantTemporaryRights();
-  SingleCollectionTransaction trx(
-      transaction::StandaloneContext::Create(_vocbase), colName,
-      AccessMode::Type::WRITE);
-  trx.addHint(transaction::Hints::Hint::RECOVERY);  // to turn off waitForSync!
-
-  Result res = trx.begin();
-
-  if (!res.ok()) {
-    errorMsg = "unable to start transaction: " + res.errorMessage();
-    res.reset(res.errorNumber(), errorMsg);
-    return res.errorNumber();
-  }
-
-  int resCode =
-      processRestoreDataBatch(trx, colName, useRevision, force, errorMsg);
-  res.reset(resCode, errorMsg);
-  res = trx.finish(res);
-
-  return res.errorNumber();
-}
-
-int restoreDataParser(char const* ptr, char const* pos,
-                      std::string const& invalidMsg, bool useRevision,
-                      std::string& errorMsg, std::string& key,
-                      VPackBuilder& builder, VPackSlice& doc,
-                      TRI_replication_operation_e& type) {
-  builder.clear();
-
-  try {
-    VPackParser parser(builder);
-    parser.parse(ptr, static_cast<size_t>(pos - ptr));
-  } catch (VPackException const&) {
-    // Could not parse the given string
-    errorMsg = invalidMsg;
-
-    return TRI_ERROR_HTTP_CORRUPTED_JSON;
-  } catch (std::exception const&) {
-    // Could not even build the string
-    errorMsg = invalidMsg;
-
-    return TRI_ERROR_HTTP_CORRUPTED_JSON;
-  } catch (...) {
-    return TRI_ERROR_INTERNAL;
-  }
-
-  VPackSlice const slice = builder.slice();
-
-  if (!slice.isObject()) {
-    errorMsg = invalidMsg;
-
-    return TRI_ERROR_HTTP_CORRUPTED_JSON;
-  }
-
-  type = REPLICATION_INVALID;
-
-  for (auto const& pair : VPackObjectIterator(slice, true)) {
-    if (!pair.key.isString()) {
-      errorMsg = invalidMsg;
-
-      return TRI_ERROR_HTTP_CORRUPTED_JSON;
-    }
-
-    std::string const attributeName = pair.key.copyString();
-
-    if (attributeName == "type") {
-      if (pair.value.isNumber()) {
-        int v = pair.value.getNumericValue<int>();
-        if (v == 2301) {
-          // pre-3.0 type for edges
-          type = REPLICATION_MARKER_DOCUMENT;
-        } else {
-          type = static_cast<TRI_replication_operation_e>(v);
-        }
-      }
-    }
-
-    else if (attributeName == "data") {
-      if (pair.value.isObject()) {
-        doc = pair.value;
-
-        if (doc.hasKey(StaticStrings::KeyString)) {
-          key = doc.get(StaticStrings::KeyString).copyString();
-        }
-      }
-    }
-
-    else if (attributeName == "key") {
-      if (key.empty()) {
-        key = pair.value.copyString();
-      }
-    }
-  }
-
-  if (type == REPLICATION_MARKER_DOCUMENT && !doc.isObject()) {
-    errorMsg = "got document marker without contents";
-    return TRI_ERROR_HTTP_BAD_PARAMETER;
-  }
-
-  if (key.empty()) {
-    errorMsg = invalidMsg;
-
-    return TRI_ERROR_HTTP_BAD_PARAMETER;
-  }
-
-  return TRI_ERROR_NO_ERROR;
-}
-
-////////////////////////////////////////////////////////////////////////////////
 /// @brief determine the chunk size
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -3195,20 +2396,6 @@ uint64_t RocksDBRestReplicationHandler::determineChunkSize() const {
 
   return chunkSize;
 }
-
-void RocksDBRestReplicationHandler::grantTemporaryRights() {
-  AuthenticationFeature* auth =
-      FeatureCacheFeature::instance()->authenticationFeature();
-  if (auth->isActive() && ExecContext::CURRENT != nullptr) {
-    AuthLevel level = auth->canUseDatabase(ExecContext::CURRENT->user(), _vocbase->name());
-    if (level == AuthLevel::RW) {
-      // If you have administrative access on this database,
-      // we grant you everything for restore.
-      ExecContext::CURRENT = nullptr;
-    }
-  }
-}
-
 
 //////////////////////////////////////////////////////////////////////////////
 /// @brief condition locker to wake up holdReadLockCollection jobs
