+++ arangod/RocksDBEngine/RocksDBRestReplicationHandler.h
@@ -25,25 +25,17 @@
 #ifndef ARANGOD_ROCKSDB_ROCKSDB_REST_REPLICATION_HANDLER_H
 #define ARANGOD_ROCKSDB_ROCKSDB_REST_REPLICATION_HANDLER_H 1
 
-#include "Basics/Common.h"
+#include "RestHandler/RestReplicationHandler.h"
 
-#include "RestHandler/RestVocbaseBaseHandler.h"
 #include "RocksDBEngine/RocksDBReplicationManager.h"
-#include "VocBase/replication-common.h"
 
 namespace arangodb {
-class ClusterInfo;
-class CollectionNameResolver;
-class LogicalCollection;
-namespace transaction {
-class Methods;
-};
 
 ////////////////////////////////////////////////////////////////////////////////
 /// @brief replication request handler
 ////////////////////////////////////////////////////////////////////////////////
 
-class RocksDBRestReplicationHandler : public RestVocbaseBaseHandler {
+class RocksDBRestReplicationHandler : public RestReplicationHandler {
  public:
   RocksDBRestReplicationHandler(GeneralRequest*, GeneralResponse*);
   ~RocksDBRestReplicationHandler();
@@ -56,12 +48,6 @@ class RocksDBRestReplicationHandler : public RestVocbaseBaseHandler {
 
  private:
   //////////////////////////////////////////////////////////////////////////////
-  /// @brief creates an error if called on a coordinator server
-  //////////////////////////////////////////////////////////////////////////////
-
-  bool isCoordinatorError();
-
-  //////////////////////////////////////////////////////////////////////////////
   /// @brief return the state of the replication logger
   /// @route GET logger-state
   /// @caller Syncer::getMasterState
@@ -71,68 +57,38 @@ class RocksDBRestReplicationHandler : public RestVocbaseBaseHandler {
   ///           * clients (list of followers)
   //////////////////////////////////////////////////////////////////////////////
 
-  void handleCommandLoggerState();
+  void handleCommandLoggerState() override;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief handle a follow command for the replication log
   //////////////////////////////////////////////////////////////////////////////
 
-  void handleCommandLoggerFollow();
+  void handleCommandLoggerFollow() override;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief handle the command to determine the transactions that were open
   /// at a certain point in time
   //////////////////////////////////////////////////////////////////////////////
 
-  void handleCommandDetermineOpenTransactions();
-
-  //////////////////////////////////////////////////////////////////////////////
-  /// @brief turn the server into a slave of another
-  //////////////////////////////////////////////////////////////////////////////
-
-  void handleCommandMakeSlave();
+  void handleCommandDetermineOpenTransactions() override;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief handle a batch command
   //////////////////////////////////////////////////////////////////////////////
 
-  void handleCommandBatch();
-
-  //////////////////////////////////////////////////////////////////////////////
-  /// @brief forward a command in the coordinator case
-  //////////////////////////////////////////////////////////////////////////////
-
-  void handleTrampolineCoordinator();
+  void handleCommandBatch() override;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief return the inventory (current replication and collection state)
   //////////////////////////////////////////////////////////////////////////////
 
-  void handleCommandInventory();
-
-  //////////////////////////////////////////////////////////////////////////////
-  /// @brief returns the cluster inventory, only on coordinator
-  //////////////////////////////////////////////////////////////////////////////
-
-  void handleCommandClusterInventory();
+  void handleCommandInventory() override;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief handle a restore command for a specific collection
   //////////////////////////////////////////////////////////////////////////////
 
-  void handleCommandRestoreCollection();
-
-  //////////////////////////////////////////////////////////////////////////////
-  /// @brief handle a restore command for a specific collection
-  //////////////////////////////////////////////////////////////////////////////
-
-  void handleCommandRestoreIndexes();
-
-  //////////////////////////////////////////////////////////////////////////////
-  /// @brief handle a restore command for a specific collection
-  //////////////////////////////////////////////////////////////////////////////
-
-  void handleCommandRestoreData();
+  void handleCommandRestoreCollection() override;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief produce list of keys for a specific collection
@@ -273,20 +229,13 @@ class RocksDBRestReplicationHandler : public RestVocbaseBaseHandler {
   /// @brief restores the indexes of a collection TODO MOVE
   //////////////////////////////////////////////////////////////////////////////
 
-  int processRestoreIndexes(VPackSlice const&, bool, std::string&);
+  int processRestoreIndexes(VPackSlice const&, bool, std::string&) override;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief restores the indexes of a collection, coordinator case
   //////////////////////////////////////////////////////////////////////////////
 
-  int processRestoreIndexesCoordinator(VPackSlice const&, bool, std::string&);
-
-  //////////////////////////////////////////////////////////////////////////////
-  /// @brief restores the data of a collection TODO MOVE
-  //////////////////////////////////////////////////////////////////////////////
-
-  int processRestoreDataBatch(transaction::Methods&, std::string const&, bool,
-                              bool, std::string&);
+  int processRestoreIndexesCoordinator(VPackSlice const&, bool, std::string&) override;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief restores the data of a collection TODO MOVE
@@ -300,12 +249,6 @@ class RocksDBRestReplicationHandler : public RestVocbaseBaseHandler {
   //////////////////////////////////////////////////////////////////////////////
 
   uint64_t determineChunkSize() const;
-
-  //////////////////////////////////////////////////////////////////////////////
-  /// @brief Grant temporary restore rights
-  //////////////////////////////////////////////////////////////////////////////
-
-  void grantTemporaryRights();
 
  private:
   //////////////////////////////////////////////////////////////////////////////
